refactor(swift): split BackendManager.swift into focused files
Extract PythonInstaller, OpencodeInstaller, ObsidianPluginManager,
ProcessUtils into separate files. Introduce BackendPaths struct for
all path/URL constants including health check URLs and port numbers.
Add Logging.swift with shared subsystem constant derived from bundle id.
