+++ DragonglassApp/Dragonglass/Logging.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+let subsystem = Bundle.main.bundleIdentifier ?? "com.lua.Dragonglass"
